This fixes a regression with `check_file_contain_within`
With [this PR](https://github.com/serverspec/specinfra/pull/191) a
regression was introduced with the `before` and `after` matchers when
they aren't being used together.

With PR 191 the following:

```
  it { should contain('rspec').before(/^end/) }
```

would have resulted in:

```
  sed -n 1,/\\^end/p Gemfile | sed -n 1,/\\^end/p | grep -qs -- rspec - || sed -n 1,/\\^end/p Gemfile | sed -n 1,/\\^end/p | grep -qFs -- rspec -
```

as well as:

```
    it { should contain('rspec').after(/^group :test do/) }
```

would result in:

```
    sed -n /\\^group\\ :test\\ do/,\\$p Gemfile | sed -n 1,\\$p | grep -qs -- rspec - || sed -n /\\^group\\ :test\\ do/,\\$p Gemfile | sed -n 1,\\$p | grep -qFs -- rspec -
```

Notice in both scenarios the `sed` command is unnecessarily duplicated.

While neither of these should actually cause an assertion to pass when
it should fail, they also do not add anything for either situation.