Do not depend on pipe buffer size
Fix for a problem mentioned in the comments for https://github.com/serverspec/specinfra/pull/442

This solution does not depend on a pipe read size and also works a bit faster then a default one.

NB: block behavior has changed in terms of output buffers updating. They are updated at the end of a command run only. However, this should not be a problem as far as I see the rest of the #spawn_command code. 

P.S. I didn't check if other backends may have similar problem.