support sticky bit of file mode on FreeBSD
On Linux, get_mode returns sticky bit if set.
On FreeBSD it should follow the behavior.

When a file named 'foo', whose mode is 0755,
`stat -f%M%Lp filepath` returns '0755' but it returns '755' on Linux.
When a file named 'foo', whose mode is 0055,
`stat -f%M%Lp filepath` returns '0055' but it returns '55' on Linux.