Use regexp to match exactly the repo name
If you have different repos with name like gvp-bi-common, gvp-bi-common-x86_64, gvp-bi-common-noarch, etc...and you want to verify if they exist or not, with the current code it fails even if the repo not exist because it matches only the beginning of the name. Matching the name exactly has the desired behavior.